% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSConvert_Deploy.R, R/MSconvertR.R
\name{MSConvert_check}
\alias{MSConvert_check}
\alias{MSConvert_Download}
\alias{MSConvert_Deploy}
\alias{msConvert}
\alias{msConvertDir}
\alias{msConvert2mzML}
\alias{msConvert2mzXML}
\title{MSConvert_check}
\usage{
MSConvert_check()

MSConvert_Download(save_path = tempdir())

MSConvert_Deploy(pwiz.bz = MSConvert_Download())

msConvert(
  raw.files,
  ms.data.names,
  format.to = "mzML",
  BPPARAM = BiocParallel::SnowParam(workers = parallel::detectCores() - 1)
)

msConvertDir(raw.path, format.to = "mzXML")

msConvert2mzML(
  raw.files,
  mzML.files,
  BPPARAM = BiocParallel::SnowParam(workers = parallel::detectCores() - 1)
)

msConvert2mzXML(
  raw.files,
  mzXML.files,
  BPPARAM = BiocParallel::SnowParam(workers = parallel::detectCores() - 1)
)
}
\description{
MSConvert_check

MSConvert_Download

MSConvert_Deploy

msConvert

msConvertDir

msConvert2mzML

msConvert2mzXML
}
\section{Functions}{
\itemize{
\item \code{MSConvert_check()}: check if MSconvert ready

\item \code{MSConvert_Download()}: MSConvert_Download

\item \code{MSConvert_Deploy()}: MSConvert_Deploy

\item \code{msConvert()}: msConvert

\item \code{msConvertDir()}: msConvertDir

\item \code{msConvert2mzML()}: msConvert2mzML

\item \code{msConvert2mzXML()}: msConvert2mzXML

}}
