% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSconvertR.R
\name{msConvert2mzML}
\alias{msConvert2mzML}
\title{Title}
\usage{
msConvert2mzML(
  raw.files,
  mzML.files,
  BPPARAM = BiocParallel::SnowParam(workers = parallel::detectCores() - 1)
)
}
\arguments{
\item{BPPARAM}{}
}
\description{
Title
}
