% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSconvertR.R
\name{msConvert2SciexMultipleWiff}
\alias{msConvert2SciexMultipleWiff}
\title{msConvert}
\usage{
msConvert2SciexMultipleWiff(
  raw.files,
  dir.to = dirname(raw.files),
  format.to = "mzML",
  BPPARAM = BiocParallel::SnowParam(workers = parallel::detectCores() - 1)
)
}
\arguments{
\item{raw.files}{data input}

\item{dir.to}{dirname(raw.files)}

\item{format.to}{"mzML}

\item{BPPARAM}{BPPARAM}
}
\description{
msConvert
}
\section{Functions}{
\itemize{
\item \code{msConvert2SciexMultipleWiff()}: Convert raw data from mass spectrometry

}}
