% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSconvertR.R
\name{MSConvert_check}
\alias{MSConvert_check}
\alias{msConvertDir}
\alias{msConvert2mzML}
\alias{MSConvert_Download}
\alias{MSConvert_Deploy}
\title{MSConvert_check}
\usage{
MSConvert_check()

msConvertDir(raw.path, format.to = "mzML")

msConvert2mzML(
  raw.files,
  mzML.files,
  BPPARAM = BiocParallel::SnowParam(workers = parallel::detectCores() - 1)
)

MSConvert_Download(save_path = tempdir())

MSConvert_Deploy(pwiz.bz = MSConvert_Download())
}
\description{
MSConvert_check

msConvertDir

msConvert2mzML

MSConvert_Download

MSConvert_Deploy
}
\section{Functions}{
\itemize{
\item \code{MSConvert_check()}: check if MSconvert ready

\item \code{msConvertDir()}: msConvertDir

\item \code{msConvert2mzML()}: msConvert2mzML

\item \code{MSConvert_Download()}: MSConvert_Download

\item \code{MSConvert_Deploy()}: MSConvert_Deploy

}}
